Use new buildProgrammingExpressionMarkdown for StyledCodeBlock
