+++ apps/src/templates/lessonOverview/StyledCodeBlock.jsx
@@ -2,6 +2,16 @@ import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import SafeMarkdown from '@cdo/apps/templates/SafeMarkdown';
 
+export const buildProgrammingExpressionMarkdown = function(
+  programmingExpression
+) {
+  let block = `\`${programmingExpression.name}\``;
+  if (programmingExpression.color) {
+    block += `(${programmingExpression.color})`;
+  }
+  return `[${block}](${programmingExpression.link})`;
+};
+
 export default class StyledCodeBlock extends Component {
   static propTypes = {
     programmingExpression: PropTypes.shape({
@@ -14,12 +24,10 @@ export default class StyledCodeBlock extends Component {
   render() {
     const {programmingExpression} = this.props;
 
-    let blockMarkdown = programmingExpression.color
-      ? `[\`${programmingExpression.name}\`(${programmingExpression.color})](${
-          programmingExpression.link
-        })`
-      : `[\`${programmingExpression.name}\`](${programmingExpression.link})`;
-
-    return <SafeMarkdown markdown={blockMarkdown} />;
+    return (
+      <SafeMarkdown
+        markdown={buildProgrammingExpressionMarkdown(programmingExpression)}
+      />
+    );
   }
 }
