+++ apps/test/unit/templates/lessonOverview/StyledCodeBlockTest.jsx
@@ -1,9 +1,32 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 import {expect} from '../../../util/reconfiguredChai';
-import StyledCodeBlock from '@cdo/apps/templates/lessonOverview/StyledCodeBlock';
+import StyledCodeBlock, {
+  buildProgrammingExpressionMarkdown
+} from '@cdo/apps/templates/lessonOverview/StyledCodeBlock';
 
 describe('StyledCodeBlock', () => {
+  describe('buildProgrammingExpressionMarkdown', () => {
+    it('builds a full visual code block markdown expression', () => {
+      const input = {
+        color: '#c0ffee',
+        link: 'https://example.com',
+        name: 'test block'
+      };
+      const expected = '[`test block`(#c0ffee)](https://example.com)';
+      expect(buildProgrammingExpressionMarkdown(input)).to.equal(expected);
+    });
+
+    it('builds a regular code block when not given a color', () => {
+      const input = {
+        link: 'https://example.com',
+        name: 'test block'
+      };
+      const expected = '[`test block`](https://example.com)';
+      expect(buildProgrammingExpressionMarkdown(input)).to.equal(expected);
+    });
+  });
+
   it('build full block markdown when color supplied', () => {
     const wrapper = shallow(
       <StyledCodeBlock
